Separar base y base de notificaciones
