+++ common/utilities/ArtchiveBase.php
@@ -8,8 +8,6 @@ use yii\helpers\Url;
 use yii\helpers\Html;
 
 use common\models\User;
-use common\models\Notificaciones;
-use common\models\TiposNotificaciones;
 
 /**
  *
@@ -22,20 +20,9 @@ class ArtchiveBase extends \yii\db\ActiveRecord
      * @return bool Por defecto guarda el historial, para que no lo guarde
      * habría que devolver falso.
      */
-    public function getGuardarHistorial()
+    public function isHistorialSaved()
     {
         return true;
-    }
-
-    /**
-     * Indica si se debe enviar una notificación al crear.
-     *
-     * @return bool Por defecto no se manda. Para que se mande habría que
-     * devolver true.
-     */
-    public function getEnviarNotificacion()
-    {
-        return false;
     }
 
     /**
@@ -189,82 +176,17 @@ class ArtchiveBase extends \yii\db\ActiveRecord
         return 'Ha eliminado ' . $this->getUnName();
     }
 
-    /**
-     * Crea el contenido de una notificación.
-     *
-     * @return string
-     */
-    public function getNotificacionContenido()
-    {
-        $name = $this->getUnName();
-        return "Has recibido $name.";
-    }
-
-    /**
-     * Devuelve el tipo de notificación basándose en el nombre de la tabla.
-     * @return string
-     */
-    public function getNotificacionTipo()
-    {
-        $nombre = strtolower(str_replace('_', ' ', static::tableName()));
-        $tipo = TiposNotificaciones::findOne(['tipo' => $nombre]);
-        return $tipo->id;
-    }
-
-    /**
-     * Devuelve el id que indica quién recibe la notificación.
-     * Por defecto es el usuario_id de la clase actual.
-     *
-     * @return int
-     */
-    public function getNotificacionReceptor()
-    {
-        return $this->usuario_id;
-    }
-
-    /**
-     * Devuelve la url de la notificación en caso de que sea diferente a la
-     * url general.
-     *
-     * @return string|bool
-     */
-    public function getNotificacionUrl()
-    {
-        return false;
-    }
-
-    /**
-     * Crea la notificación.
-     *
-     * @return bool Verdadero si se ha creado satisfactoriamente y falso si no.
-     */
-    public function createNotificacion()
-    {
-        $contenido = Html::a(
-            $this->getNotificacionContenido(),
-            $this->getNotificacionUrl() ?: $this->getRawUrl()
-        );
-        $noti = new Notificaciones([
-            'notificacion' => $contenido,
-            'usuario_id' => $this->getNotificacionReceptor(),
-            'tipo_notificacion_id' => $this->getNotificacionTipo(),
-        ]);
-        return $noti->save();
-    }
-
     public function afterSave($insert, $changedAttributes)
     {
         parent::afterSave($insert, $changedAttributes);
-        if ($this->getGuardarHistorial()) {
+        if ($this->isHistorialSaved()) {
             if ($insert) {
                 Historial::crearHistorial($this->getInsertMessage(), $this->getRawUrl());
             } else {
                 Historial::crearHistorial($this->getUpdateMessage(), $this->getRawUrl());
             }
         }
-        if ($this->getEnviarNotificacion() && $insert) {
-            $this->createNotificacion();
-        }
+
         return true;
     }
 
@@ -274,7 +196,7 @@ class ArtchiveBase extends \yii\db\ActiveRecord
             return false;
         }
         $data = isset($this->{$this->getDataName()}) ? ': "' . $this->{$this->getDataName()} . '"' : '';
-        if ($this->getGuardarHistorial()) {
+        if ($this->isHistorialSaved()) {
             Historial::crearHistorial($this->getDeleteMessage() . "$data.", false);
         }
         return true;
